Run one process for each file
With this change, we spawn a process for each file on the file system.
This process is responsible for compiling and rendering the file, and
for maintaing a list o other file processes that depend on it. This
implementation of subscribers is rudimentary because it doesn't allow
removing subscribers. In the following changes we have to fix that, and
update the watcher to use that information in order to only trigger
recompiles for the files that need it.