Rewrite watcher and compiler
This ended up being a bigger change than I initially planned. Unfortunately, to get some things working, I had to touch almost everything. The main feature in this PR is a graph of dependencies between files, which means that whenever a file changes, we can only recompile the parts that need it.

A FileProcess represents each file that's compiled independently of another. This process is named using the file's relative path and managed in a new DynamicSupervisor called FileRegistry.

With this change, everything becomes a file! For instance, when a template _index.slime_ has a `layout: _layout.slime`, each file has a different process, and there is a dependency between them. This was a significant change that required moving around a lot of pieces in the compiler code.

The hardest part was, without a doubt, passing the current `file` to the included file in the `include` helper. This one leveled up my meta programming skills and made me go through the source code of Slime, EEx, and PhoenixSlime for hours. My findings are that these templating engines are not designed to be used nicely in the way our scenario, but it works. So, whenever we have to render a Slime template, we have to define a module in runtime, compile it, and then call a function.