Restructure the code
Why:

* This has actually been nagging me for a while because everything is so
inconsistently placed all over the codebase, it actually makes it
difficult to extend.
* Collections are only used for the compilation process, but they're not
part of the compiler.
* Context has the same scenario plus the fact that it is actually only a
compilation context. The very concept of it is keeping track of
variables for compilation.
* FileProcess and FileRegistry do not only represent files but folders
as well, being confusing and not aptly named.
* FileProcess and FileRegistry only exist effectively do parallel
compilation, but they're outside the compiler context. Even though they
have full knowledge of the logic of that context.
* FileProcess and FileRegistry handle the same concepts and work
depending on each other but are treated as two different top-level
modules, as if they were independent. The dependency is not obvious for
newcommers to understand they're dependent. They actually seem related
to the File module which is not true.

This change addresses the need by:

* Moving Collections and Context to the Compiler context.
* Moving FileProcess and FileRegistry to the Compiler context,
namespacing them under a Parallel module and renaming them to Node and
Registry, respectively (I'm not 100% sold on Node since it is a generic
module + technically folders are files in Unix systems, albeit this
feels like a very specific technical detail).
* Creating a new Compiler supervisor to abstract the details of what
happens inside the compilation step.