Only refresh browsers after the compilation is completed
There's a timing issue when the browser refreshes. For instance, let's say we have an `about.slime` that includes ` _header.slime`. When the `_header.slime` changes, we compile it's and asynchronously tell the `about.slime` to recompile. After that, we tell the browser to refresh. The issue is that it doesn't wait for the `about.slime` to recompile before refreshing the browser.

In this PR, we move the subscription logic from the watcher into a new module `BrowserSubscribers` and the existing `Node`. The sockets subscribe to the `BrowserSubscribers`, and the `Node` sends a notify message after it compiles. There's a throttle mechanism so that the `BrowserSubscribers` only refreshes the browsers if it doesn't receive any new notify message for 400ms.