Run preprocessors in sequence
This change is the first step towards supporting custom preprocessors. In this change, we add support for preprocessor chains, which allows us to support more than HTML preprocessors, and to reuse functionality between them. For instance, we can chain preprocessors in a way that will enable CSS files to have frontmatter data and EEx interpolation.