Eagerly load modules to determine view helpers
Why:

* Applications and projects that implemented custom view helpers
depended on Still.
* Thus when the async compilation started, the applications weren't
loaded and we were unable to import those custom view helpers.

This change addresses the need by:

* Grabbing all compiled modules
* Filtering the Elixir ones (implement `module_info/1`). This causes all
compiled modules to be loaded (which is a big shame) but we are not able
to do it any other way with the current approach, as
`Kernel.function_exported?/3` and `Kernel.macro_exported?/3` return
false since these modules are not yet loaded.
* Importing all necessary modules.