Define compilation timeout
Why:

* I was recently faced with an issue when my computer was burning up and
running VERY slowly: when compiling the project, compilation would
timeout due to GenServer's implicit `5_000`ms timeout.
* This timeout isn't necessarily a bad thing as it might developers face
some issues. However, it might also be a pain for larger projects in
computers with less resources.

This change addresses the need by:

* THIS CODE IS NOT MEANT TO BE A SOLUTION
* I'm looking to set up the discussion: should we remove the timeout?
Should we set a larger value? It happened to me quite often so maybe
this isn't the best value. Should we make it configurable? If so, with
what default value?