Error overlay
This is a mechanism similar to the one you have on a React app: when there's a compilation error or an exception, you get a screen showing more details on that error. With this change, ours will look like this:

![Print of the error overlay](https://user-images.githubusercontent.com/934580/94999001-0c13ba80-05ae-11eb-8e75-27644a51d880.png)

When the error is in a CSS file we can't display the context and stacktrace.

How does it work? The first thing I did was to prevent the app from crashing when there's a compilation error. So first we capture `CompileError` and turn it into a `ProcessorError` where we add more information. Then we capture that new error on the `Node` module.

When a file is being `rendered` inside another file, we return a template that will display an overlay. For instance, for HTML files, when we get the error, we create an HTML snippet that displays the error. For JS we take that same HTML and return javascript that will change the body to that HTML. For CSS we change the `content` of the pseudo-element `body::after`.

When a file is being `compiled`, we do something similar to what we do on the `render`, but instead we wrap that content in the dev layout (when appropriate) and write that file to the place where the failing file was expected.

It's a simple mechanism that gives immediate feedback to the user. Plus, because we have auto-refresh set in the development layout when the error is fixed, the browser refreshes, and the overlay error disappears automatically.