Add Still's website
With this change, we use Still to build Still's website. I'm moving the documentation to the website because it allows us to be more expressive. There are a few other changes to the documentation which I believe are improvements, but there is so much more to do. There's also the first version of an announcement blog post, but once again, still being worked on, I'm happy to remove everything and start over.

[Link the website](https://subvisual.github.io/still).