User View Helpers and Docs
With this change, we replace the existing mechanism to load the user's view helpers. Up until now, the modules would be imported if they implemented a specific behavior. Unfortunately, this mechanism slows down the compilation, and it's noticeable on our small website. This may be our fault for lacking the knowledge to come up with a better solution, but for now, simply having the users declare in the config file which modules are the view helpers saves us this trouble.

This change fixes https://github.com/subvisual/still/issues/41