URL Fingerprinting
URL fingerprinting is important for caching in production. In this change we move the code to calculate the output path to a new preprocessor, and we add another preprocessor to add the fingerprint to the output path. There's also a config to enable this feature when running the task `compile`.

There are also two new helpers to link to css and js files built on top of this.