Remove Still.Preprocessor.SyntaxError
In 71c4af5089cc13352a7df932abb640092bb62b25 we created a new mechanism
to handle errors in the preprocessors: instead of exiting the
application, we catch the exceptions, turn them into
%PreprocessorError{} and then turn those exceptions into an error
overlay that we show to the users. Still, some errors are being handled
by the previous mechanism that turns errors into SyntaxError and logs
them to the console. In this change, we remove this old mechanism so that
every error goes straight to the error overlay, informing the user.