Add archives to repo
With this change we can install the archive like this:

```
mix archive.install https://raw.githubusercontent.com/subvisual/still/master/archives/still.ez
```

It took me a while to figure out, but our archives had all the mix tasks, including `still.compile` and `still.dev`. Because of that, when I had the archive installed, and I ran `mix still.compile`, it was using the task defined in the archive instead of the one in the dep. I went through phoenix's codebase and this is how they solved: a nested library just for the generator.

The documentation will be updated in a different pull request.