Template variables are now module variables
With this change, every variable is available as a module variable instead of as a function variable.

This template:

```
---
title: "Titlte"
---
h1 = title
```

Becomes this

```
---
title: "Titlte"
---
h1 = @title
```

This is what happens in Phoenix, and I think I know why. Storytime!

---

There's a blog, and each entry is like this:

```
---
layout: _layout.slime
---
h1 Some title.
```

And the layout is like this:

```
.nav
  a Link1
  a Link2

= children
```

But in some pages, we want to hide the nav:

```
---
layout: _layout.slime
hide_anv: true
---
h1 Some title.
```

And on the layout we should be able to do something like this:

```
= if !hide_nav do
  .nav
    a Link1
    a Link2

= children
```

Unfortunately, it doesn't because for the entries that don't hide the nav the variable is not set. So I could either add it to every entry, which is a mess or define a default value:

```
---
hide_layout: false
---
= if !hide_nav do
  .nav
    a Link1
    a Link2

= children
```

But this solution doesn't work for complex variables like maps. You can't set a default value for a map (you can, but it's a mess). That's where the `assigns` map you have in Phoenix comes in:


```
= if Map.get(assigns, :hide_nav, false) do
  .nav
    a Link1
    a Link2

= children
```

And that's why I'm making this change. I don't know if Phoenix works this way for this reason, but it does solve my problems.