Add CompilationStage
The _CompilationStage_ is a new mechanism that compiles, and recompiles, files:

* the `Traverse` goes through the file system adding files to the _CompilationStage_.
* the `Watcher` adds changed files to the _CompilationStage_.
* when the `Node` recompiles, it adds its dependencies to the _CompilationStage_.

Because the _CompilationStage_ knows which files are pending compilation, we use it o inform the browser to refresh, or the compilation task to complete. Unfortunately, the compilation should be a little slower, but I cannot think of another solution right now.