Improved error overlay
I noticed that the error overlay couldn't render some errors, so I started looking into it. The stack trace sometimes has the arity on the third argument, but sometimes has the actual arguments. I fixed that, and then I noticed that there are whitespaces all over the place in the context section. Unfortunately, there's no easy way to remove those without parsing the HTML, so I added floki for that.

It  goes from this:

![Screen Shot 2020-12-13 at 16 38 53](https://user-images.githubusercontent.com/934580/102018015-7ab98500-3d62-11eb-98b8-1639ee8129a7.png)

To this:

![Screen Shot 2020-12-13 at 16 42 47](https://user-images.githubusercontent.com/934580/102018017-7beab200-3d62-11eb-87c6-ef6e6b5c23fa.png)