Everything is a preprocessor!
This change goes towards our initial intention of making everything a preprocessor. With this change, we can combine the preprocessors to any scenario. For instance, on a new image processing preprocessor I'm taking out the "AddContent" preprocessor because I don't want to read the content in Elixir, I want to send the file directly to Rust where it will be read as a stream. In the same scenario, I don't want to save the file, because it will be saved by rust. I think this design really is a step forward.