Image Preprocessor
This change introduces the **optional** image preprocessor built on top of [`imageflow_ex`](https://github.com/naps62/imageflow_ex) which defaults to ["mogrify"](https://github.com/route/mogrify) which uses ImageMagick.
The API is still low-level, but it's a start. 

The preprocessor can generate multiple output files and apply transformations to them. 

Image processing is pretty slow, so we use the files's modified time and name to decide whether to run or not: if the output files already exist and the input file's modified time is before, or the same, as the output files', it doesn't run.

The transformations map is hashed and encoded in the file's name, so we know that if we change the transformation, the expected output name will also change, and the preprocessor will run.

Besides the preprocessor, there's also a `ResponsiveImage` view helper that sets the sizes automatically based on the input file's width. It's a pretty dump setup for now.

Extra:

* `variables` was renamed to `metadata` because it was starting to get confusing. I find that metadata is a better name.
* the compilation timeout was increased by 5 seconds before processing images is a slow operation and sometimes it times out. Keep in mind that this operation is only slow when we are building the site for the first time or when we remove the output folder.