Add Rust as dependency
It's currently impossible to compile the project without a recent
version of Rust. This PR adds it as an explicit dependency by including
it in the `.tool-versions` file and in the `bin/setup` script.

`bin/setup` doesn't go deep into installing Rust as we're doing the best
we can to remove it as a dependency.