Implement image adapters
We're currently having issues when developing Still with the Rust
dependency. Although it's not an issue for users, it makes the
development experience much worse.

This PR removes the dependency from Imageflow by creating the concept of
an image adapter and setting the Mogrify implementation to be the
default one.

It comes hand in hand with the creation of
[`still_imageflow`](https://github.com/still-ex/still_imageflow) which
contains the now-removed Imageflow module.

Any user that wishes to make use of Imageflow, should instead include
`still_imageflow` in the dependencies and configure their project to set
it as the correct adapter.