Fix Watcher crashing when removing a file
When a file is removed, the watcher is crashing due to the `Registry`
call to `DynamicSupervisor.terminate_child/2`.

The second argument is `nil` because it attempts to get the PID by
running `Process.whereis(file |> String.to_atom())`. However, the
registered name is a string.

Since this call crashes due to the lack of a function clause, the
watcher would crash as well.