Better errors using the built-in Exception module
I found https://hexdocs.pm/elixir/Exception.html when I was going through Phoenix's source code. With this, we can remove some of the code we have to format errors and format errors in a familiar way to the users. I also moved some things around to ensure we only have to format errors in one place, and we can support any kind of error.