Raise when importing files that don't exist
In the current implementation, we can import a file that doesn't exist. We'll create a Node for it and then the `AddContent` preprocessor will throw, creating an error in the ErrorCache for a file that doesn't exist. This error is impossible to clear, so we would have to stop the server and start it again. Including a file that doesn't exist should throw an error in the processor including the file.