Find preprocessors using a regex
This change allows finding the preprocessor chain using a regex instead of just the file extension.

Using the file extensions is fine for simple use cases, but it's not flexible enough. It doesn't make sense to force users to use a custom extension when they need to have two compilation chains for the same file type.