Rewrite preprocessors (#142)
At the moment, they way we execute a preprocessor chain goes something like this:

```elixir
def run(source_file, [preprocessor | remaining_preprocessors]) do
  preprocessor.run(source_file)
  |> run(remaining_preprocessors)
end
```

We execute one after the other, using the output of one as the input of the next.

This has two relevant limitations: 1) a preprocessor cannot stop the execution of the chain, and 2) a preprocessor cannot execute after the remaining preprocessor. The profiler could be a preprocessor that stores the current time, calls the next preprocessor, and when the next returns, calculates the diff. The pass-through copy could be a preprocessor that checks whether a file is eligible for copy and if it is, stops the chain right there. There are many more use cases for such functionality.

With this change, we rewrite the internals of Preprocessor so that each preprocessor calls the next. This change doesn't affect the existing API, instead, it extends it.

* `render/1` can return a `SourceFile` but can also return `{:cont, source_file}` and `{:halt, source_file}` in order to control whether the chain should continue its execution or not.
* a new `after_render/1` callback exists to execute code after the next preprocessors finish.
* `run/1` is now `run/2` and it receives the next preprocessors in the second argument.
* `extension/1` was dropped because it isn't really necessary. I know it's a breaking change, but we might as well just remove it since almost no one is using Still **yet**.