Fix image resize and an option for quality
There's a bug in the call to resize images. The current implementation uses "xVALUE", but according to the docs, the syntax to resizes the image by width is "VALUE". The resulting images were huge because the syntax "xVALUE" resizes by height. [Check the docs for more information](https://imagemagick.org/script/command-line-processing.php#geometry).

I also took the opportunity to introduce an option to change the images' quality.