Missing dependency_chain when an error occurs outside a Preprocessor (#154)
This change addresses the issue raised in https://github.com/still-ex/still/issues/153

When you define a custom preprocessor chain, if you don't alias a module, the module won't be found, and Elixir will throw an error. This error is hidden in the console while it should have appeared on the page.

Unfortunately, for errors that happened outside a preprocessor are caught in `Still.Compiler.Incremental.Node` and transformed into a `%PreprocessorError`, the `dependency_chain` was missing, and the error overlay could not generate a proper id for the error, therefore never showing the error to the user.

On top of that, if you start the dev server with an empty project, and the error happens in your `index.html`, you also never get to see the error, you only see that the page doesn't exist. Changing the router to render the dev layout when `index.html` doesn't exist solves that.