Recompile on request
The initial goal of this change was to automatically recompile our Elixir modules on a page reload, like Phoenix. Unfortunately, recompiling the modules is not enough, and we have to "touch" the templates that use those modules to see the changes. With this insight, I decided to change things up and remake the whole compilation mechanism. Instead of recompiling templates on file change, we'll recompile them when there's a browser request. This change doesn't affect the way we write a site in any way, and it turns out that this solution is much simpler than what we had, and we no longer have to keep track of the dependencies between files and all of that.

Back to the original plan, I copied some code from Phoenix that recompiles the modules on request and put it in a plug. I also added a registry to keep track of which input files generate which output files to recompile them when they are asked for.

There's also a new `:run_type` called `:compile_metada` used to inform the preprocessor that we don't really want to compile the file, and we only care about its metadata. Each preprocessor handles this type differently, but it does save a lot of time. When we start the dev server, we compile the metadata for every file to know which files to compile when the browser asks. 

Along the way, there were other more minor issues that I found and fixed:

* the `base_url` should be a relative path that's only relevant when the site is not deployed in the root of a domain.
* the `on_compile` hook lives in the root module now. This makes it easier for us to change the implementation without breaking the API.
* when a preprocessor module doesn't exist or its run function isn't defined, it throws a proper error.