Replace still_snowpack and others with watchers (#164)
After [`still_snowpack`](https://github.com/still-ex/still_snowpack), I wanted to write a plugin to integrate with Tailwind. I spent a lot of time thinking about this, about the ergonomics and how other projects did it, and I believe we are going in the wrong direction. The setup for `still_snowpack` is confusing and most of it could be accomplished through the CLI. And the same applies to Tailwind, where running the binary would be more than enough.

There are scenarios where it makes sense to build on top of [`still_node`](https://github.com/still-ex/still_node), but that shouldn't be the only solution. This is where `watchers` come in (it was literally copy-pasted from Phoenix). You can see on the example I wrote for tailwind that with just one config you have tailwind up and running and integrated with Still.

On top of that, the `ignored_file` config allows us to have the Node projects inside the input folder, next to the rest of the code.