Add preprocessor to replace HTML images with responsive images (#167)
With this change, we introduce a preprocessor that's capable of replacing images with responsive images in HTML. Images that have a `srcset` or a `no-responsive-image` are ignored.

The markdown preprocessor was updated to run the generated HTML through this new preprocessor, making it super simple to use for blogs. This behavior is controlled through a flag.

Extra:

The calls to `ErrorCache` are not async. There's no reason for them to be sync that I can think of. On top of that, it was breaking the tests because the `ErrorCache` process is not running.