Better image support (#174)
This started as an attempt to replace the markdown parser with Earmark, which is built on Elixir. I noticed that Earmark supports processors, and I could use that to build responsive images from markdown. I was using an HTML parser that ran after the markdown parser. While moving the code around, I started refactoring some of the functions and moving the code related to images to `Still.Image`. There's still a coupling from the other parts of the system to `Still.Image`, but it's small and not worth removing right now.